Bypass ARP (version 2) by modifying GNET handler
Changed from previous approach where I was bypassing ARP by writing
broadcast MAC in the IP function. Now this happens in the GNET level.
The GNET handler checks for the broadcast IP destination address in the
incoming packet, if it is found then it writes to the packet the
broadcast mac address.